feat: secured the endpoints where necessary

+++ address_book/address/views.py
@@ -1,4 +1,4 @@
-from rest_framework import generics, views as rest_views, response, status
+from rest_framework import generics, views as rest_views, response, status, permissions
 
 from .models import Address
 from .serializers import AddressSerializer, AddressBatchDeletionRequestSerializer
@@ -9,6 +9,10 @@ class AddressDetail(generics.RetrieveUpdateDestroyAPIView):
 
     queryset = Address.objects.all()
     serializer_class = AddressSerializer
+    permission_classes = [permissions.IsAuthenticated]
+
+    def get_queryset(self):
+        return Address.objects.filter(owner=self.request.user)
 
 
 class AddressList(generics.ListCreateAPIView):
@@ -16,6 +20,10 @@ class AddressList(generics.ListCreateAPIView):
 
     queryset = Address.objects.all()
     serializer_class = AddressSerializer
+    permission_classes = [permissions.IsAuthenticated]
+
+    def get_queryset(self):
+        return Address.objects.filter(owner=self.request.user)
 
     def perform_create(self, serializer):
         serializer.save(owner=self.request.user)
@@ -46,6 +54,6 @@ class AddressBatchDeletion(rest_views.APIView):
         serializer = AddressBatchDeletionRequestSerializer(request.data)
         ids = serializer.data["ids"]
 
-        Address.objects.filter(id__in=ids).delete()
+        Address.objects.filter(owner=request.user, id__in=ids).delete()
 
         return response.Response(status=status.HTTP_204_NO_CONTENT)
